cleanup app and fill routes
